+++ backend/src/auth/auth.module.ts
@@ -1,8 +1,11 @@
-import { Module } from '@nestjs/common';
+import { forwardRef, Module } from '@nestjs/common';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
+import { MailModule } from '../mail/mail.module';
+import { UserModule } from '../user/user.module';
 
 @Module({
+    imports: [MailModule, forwardRef(() => UserModule)],
     controllers: [AuthController],
     providers: [AuthService],
     exports: [AuthService],