+++ backend/src/role/role.module.ts
@@ -3,7 +3,7 @@ import { RoleService } from './role.service';
 import { RoleController } from './role.controller';
 import { SequelizeModule } from '@nestjs/sequelize';
 import { RoleModel } from './models/role.model';
-import { UserModel } from '../users/models/user.model';
+import { UserModel } from '../user/models/user.model';
 import { UserRolesModel } from './models/user-roles.model';
 
 @Module({